Box2D Physics v0.0.6.4
- Physikengine Box2D implementiert
  -> Panzer um 360° drehbar
- 2 verschiedene Feuermodi hinzugefügt:
  -> 1 um auf Panzer zu schießen
  -> 2 um auf Flugzeuge zu schießen
- Gegnerbewegung abgerundet (flüssigere Bewegung)
- neue Bodentextur
- Bugfix wo PzIV keinen Turm hatte
- Bugfix bei Explosions-Animation

+++ ui/frmtanks.h
@@ -42,7 +42,7 @@ private:
     double softTerrRes;
     double hardTerrRes;
     int vehicleID;
-    double calcSpeed(int value);
+    double calcSpeed(double value);
     bool heal;
     int vel;
     int camo;
